multi: Handle chain ntfn callback in server.
This further decouples the block manager from the server by moving the
chain callback code from the block manager to server where it more
naturally belongs since it is not directly related to sync, rather it is
in response to it.

The following is a high level overview of the changes:
- Move blockchain notification callback to server
  - Use instance vars and methods directly on server
- Move funcs and consts only used by callback from blockmanager.go to server.go
- Remove no longer needed methods from peerNotifier interface
  - RelayInventory
  - TransactionConfirmed
- Remove no longer needed fields from blockManagerConfig struct
  - FeeEstimator
  - BgBlkTmplGenerator
  - NotifyWinningTickets
  - PruneRebroadcastInventory
- Move lottery data duplicate filter state to server
- Move block announce duplicate filter state to server
- Misc consistency nits

This is a part of the overall effort to decouple the block manager from
the server so it can be split out into a separate internal netsync
package.
